refactor: :truck: moved folder enums -> Language
