+++ src/project/dto/update-project.input.ts
@@ -1,6 +1,6 @@
 import { CreateProjectInput } from './create-project.input';
 import { InputType, Field, Int, PartialType } from '@nestjs/graphql';
-import { Language } from 'src/Language';
+import { Language } from 'src/enums/Language';
 
 @InputType()
 export class UpdateProjectInput extends PartialType(CreateProjectInput) {
